Allow basic capturing
With the exception of pawns, pieces can now capture. Capturing, and even
moving in general, does not take into account that one piece may block
another and therefore prevent the move. Currently, as long as the type
of movement between two positions is legal for a piece, the board will
allow a player to move a particular piece, even if intermediary pieces
are otherwise blocking the movement.